third_party/blink/web_tests/touchadjustment/html-label-expected.txt

Tests if labels are treated as clickable if the input they control is.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing small direct hits.
PASS adjusted node was A#mylink.
PASS adjusted node was LABEL#mylabel.
PASS adjusted node was INPUT#myinput.
Testing indirect hits.
PASS adjusted node was A#mylink.
FAIL adjusted node should be LABEL#mylabel. Was A#mylink.
PASS adjusted node was INPUT#myinput.
PASS successfullyParsed is true

TEST COMPLETE
